+++ utils.py
@@ -38,3 +38,9 @@ def logsumexp(arg):
             return result
     else:
         raise ValueError('argument is not an array')
+
+def hms(secs):
+    h = secs//3600
+    m = secs//60 - 60*h
+    s = (secs - 60*m - 3600*h)//1
+    return f'{h} h {m} m {s} s'
